# chore: improve jest speed on both local and CI