--- a/scripts/jest/src/environment.js
+++ b/scripts/jest/src/environment.js
@@ -1,3 +1,0 @@
-const isCI = Boolean(process.env.TF_BUILD);
-
-exports.isCI = isCI;