<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

We use most optimal setting for [`maxWorkers`](https://jestjs.io/docs/configuration#maxworkers-number--string) tested on various environments to gather the most optimal speed

### whole repo

> `lage test`

| Before | After |
|--------|--------|
| 15m 52s | 13m 14s / **𝚫 17% FASTER** |

### v9 libs

> `nx run-many -t test -p tag:vNext --exclude 'tag:tools,apps/**'` 

| Before | After |
|--------|--------|
| 7m 15s | 4m 25s / **𝚫 40% FASTER** |

### v8 libs 

> `nx run-many -t test -p tag:v8 --exclude 'tag:vNext,apps/**'`

| Before | After |
|--------|--------|
| 2m 39s | 2m 16s / **𝚫 15% FASTER** |

### v0 libs

> `nx run-many -t test -p tag:react-northstar --exclude 'tag:vNext,apps/**'`

| Before | After |
|--------|--------|
| 4m 44s | 4m 27s / **𝚫 6% FASTER** |

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Follows 
  - https://github.com/microsoft/fluentui/pull/27680
  - https://github.com/microsoft/fluentui/issues/27359