@fabricteam 2024-07-24
## 📊 Bundle size report

✅ No changes found
<div id="bundleSizeComment9423"/>